--fix=fix script and readme
